pyproject.toml
[project]
name = "rmbg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.35.0",
    "kornia>=0.8.1",
    "matplotlib>=3.10.3",
    "pillow>=11.3.0",
    "timm>=1.0.16",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.53.1",
]

